uni_res_man/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==4.4.0
django-filter==24.3
django-oauth-toolkit==3.0.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
